requirement.txt
tensorflow-gpu==1.15
keras==2.3.1
matplotlib
